pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
django-crispy-forms = "^1.11.2"
Pillow = "^8.2.0"
PyJWT = "^2.0.1"
six = "^1.16.0"
django-environ = "^0.4.5"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
